:lipstick: Delete unnecesary lines

+++ src/core/generator/noiseFunctions/perlinNoise.cpp
@@ -6,12 +6,7 @@
  */
 float PerlinNoise::interpolate(float a0, float a1, float w)
 {
-    /* // You may want clamping by inserting:
-     * if (0.0 > w) return a0;
-     * if (1.0 < w) return a1;
-     */
-    // return (a1 - a0) * w + a0;
-    // // Use this cubic interpolation [[Smoothstep]] instead, for a smooth appearance:
+    // // Use this cubic interpolation [[Smoothstep]], for a smooth appearance:
     // return (a1 - a0) * (3.0 - w * 2.0) * w * w + a0;
 
     // Use [[Smootherstep]] for an even smoother result with a second derivative equal to zero on boundaries:
@@ -33,8 +28,8 @@ vector2 PerlinNoise::randomGradient(int ix, int iy)
     a *= 2048419325;
     float random = (float)(a * (3.14159265 / ~(~0u >> 1))); // in [0, 2*Pi]
     vector2 v;
-    v.x = cos(random);
-    v.y = sin(random);
+    v.x = (float)cos(random);
+    v.y = (float)sin(random);
     return v;
 }
